+++ changelog.cpp
@@ -11,7 +11,8 @@ changelog::changelog(QWidget *parent) :
     connect(ui->cmbVersions, &QComboBox::currentTextChanged,
             this, &changelog::onCmbChange);
 
-    QFile changelog(QApplication::applicationDirPath() + "/CHANGELOG.md");
+    //QFile changelog(QApplication::applicationDirPath() + "/CHANGELOG.md");
+    QFile changelog(":/CHANGELOG.md");
     changelog.open(QIODevice::ReadOnly);
 
     QTextStream in(&changelog);
@@ -31,7 +32,7 @@ changelog::~changelog()
 void changelog::onCmbChange(const QString &selectedText)
 {
     ui->lblText->clear();
-    QFile changelog(QApplication::applicationDirPath() + "/CHANGELOG.md");
+    QFile changelog(":/CHANGELOG.md");
     changelog.open(QIODevice::ReadOnly);
 
     ui->lblVersion->setText(selectedText);
@@ -39,25 +40,26 @@ void changelog::onCmbChange(const QString &selectedText)
     QTextStream in(&changelog);
 
     bool archiveRow = false;
+    QStringList sl;
 
     while (!in.atEnd()) {
         QString line = in.readLine();
-        if (line.toUpper().contains("UNRELEASED")) continue;
+        if (line.toUpper().contains("UNRELEASED") || line.isNull() || line.isEmpty()) continue;
 
         if ( line.left(4) == "## [" && archiveRow)
             break;
 
         if ( archiveRow ) {
-            QString oldText = ui->lblText->text() + "<br>";
+//            QString oldText = ui->lblText->text() + "<br>";
             if (line.left(3) == "###") {
                 line.replace("### ", "<span><b><i>").append("</i></b></span>");
             }
             if (line.left(2) == "- ") {
                 line.append("</tt>");
             }
-            ui->lblText->setText(oldText + "<tt>" + line);
+//            ui->lblText->setText(oldText + "<tt>" + line);
+            sl.append("<tt>" + line);
         }
-
 
         if ( line.mid(4, line.indexOf("]") - 4) == selectedText ) {
             archiveRow = true;
@@ -67,6 +69,7 @@ void changelog::onCmbChange(const QString &selectedText)
             ui->lblDate->setText(newStringDate);
         }
     }
+    ui->lblText->setText(sl.join("<br>"));
 }
 
 void changelog::on_btnClose_released()