+++ mainwindow.cpp
@@ -45,6 +45,8 @@ MainWindow::MainWindow(QWidget *parent) :
     // listening on port signal from sessionManager goes to Main to update ico
     connect(sessionManager, &tsuManager::listenerUpdate, this, &MainWindow::listenerUpdate);
 
+    connect(sessionManager, &tsuManager::sessionStatisticUpdate, this, &MainWindow::sessionStatisticUpdate);
+
     // requested cancel from downloadPage (emitted from a tsucard) goes to sessionManager to remove torrent from libtorrent
     connect(downloadPage, SIGNAL(sendRequestedCancelToSession(std::string,bool)), sessionManager, SLOT(getCancelRequest(std::string,bool)));//, Qt::QueuedConnection);
 
@@ -54,17 +56,17 @@ MainWindow::MainWindow(QWidget *parent) :
     // requested resume from downloadPage (emitted from a tsucard) goes to sessionManager to resume torrent
     connect(downloadPage, SIGNAL(sendRequestedResumeToSession(std::string)), sessionManager, SLOT(getResumeRequest(std::string)));//, Qt::QueuedConnection);
 
-    // update from downloadPage to statusbar text
-    connect(downloadPage, SIGNAL(sendUpdateToStatusBar(const QString &)), this, SLOT(updateStatusBarStatistics(const QString &)));//, Qt::QueuedConnection);
+    // update from downloadPage to statusbar text DA TOGLIERE
+//    connect(downloadPage, SIGNAL(sendUpdateToStatusBar(const QString &)), this, SLOT(updateStatusBarStatistics(const QString &)));//, Qt::QueuedConnection);
 
-    // update from downloadPage to statistics page
-    connect(downloadPage, SIGNAL(sendStatisticsUpdate(const QPair<int,int> &)), statisticsPage, SLOT(updateStats(const QPair<int,int> &)));//, Qt::QueuedConnection);
+    // update from downloadPage to statistics page DA TOGLIERE
+//    connect(downloadPage, SIGNAL(sendStatisticsUpdate(const QPair<int,int> &)), statisticsPage, SLOT(updateStats(const QPair<int,int> &)));//, Qt::QueuedConnection);
 
     // popup message from downloadPage
     connect(downloadPage, SIGNAL(sendPopupInfo(QString)), this, SLOT(popupInfo(QString)));//, Qt::QueuedConnection);
 
-    // update from downloadPage to gauge
-    connect(downloadPage, SIGNAL(sendUpdateGauge(double, double)), this, SLOT(updateGauge(double, double)));//, Qt::QueuedConnection);
+    // update from downloadPage to gauge DA TOGLIERE
+//    connect(downloadPage, SIGNAL(sendUpdateGauge(double, double)), this, SLOT(updateGauge(double, double)));//, Qt::QueuedConnection);
 
     // update statusBar from downloadPage
     connect(downloadPage, SIGNAL(sendMessageToStatusBar(const QString &)), this, SLOT(updateStatusBarLabel(const QString &)));
@@ -116,11 +118,21 @@ void MainWindow::updateStatusBarLabel(const QString &msg)
     statusBar()->showMessage(msg);
 }
 
-void MainWindow::updateGauge(const double &downValue, const double &upValue)
+void MainWindow::updateGauge(const float &downValue, const float &upValue)
 {
-//    p_scaleGauge->setValue(value);
-    p_downNeedle->setCurrentValue(downValue);
-    p_upNeedle->setCurrentValue(upValue);
+    if (downValue > p_gaugeDownNeedle->getMaxValue()) {
+        p_gaugeValuesItem->setMaxValue(downValue);
+        p_gaugeDegreesItem->setMaxValue(downValue);
+        p_gaugeDownNeedle->setMaxValue(downValue);
+    }
+
+    if (upValue > p_gaugeUpNeedle->getMaxValue()) {
+        p_gaugeUpNeedle->setMaxValue(upValue);
+    }
+
+    p_gaugeSpeedLabel->setText(QString::number(downValue, 'f', 2));
+    p_gaugeDownNeedle->setCurrentValue(downValue);
+    p_gaugeUpNeedle->setCurrentValue(upValue);
 }
 
 void MainWindow::initializeScreen() {
@@ -168,10 +180,7 @@ void MainWindow::initializeScreen() {
     ui->content->addWidget(statisticsPage);
     ui->content->addWidget(downloadPage);
 
-//    p_scaleGauge = new QScale(this);
-//    p_scaleGauge->setInvertedAppearance(true);
-//    ui->contentGauge->addWidget(p_scaleGauge);
-    p_speedGauge = new QcGaugeWidget;
+    p_gauge = new QcGaugeWidget;
 //    p_SpeedGauge->addBackground(99);
 
 //    QcBackgroundItem *bkg1 = p_SpeedGauge->addBackground(92);
@@ -184,31 +193,35 @@ void MainWindow::initializeScreen() {
 //    bkg2->addColor(0.1f,Qt::gray);
 //    bkg2->addColor(1.0,Qt::darkGray);
 
-    p_speedGauge->addArc(75);
-    p_speedGauge->addDegrees(75)->setValueRange(0,100);
+    p_gauge->addArc(75);
+    p_gaugeDegreesItem = p_gauge->addDegrees(75);
+    p_gaugeDegreesItem->setValueRange(0,20);
 
-    p_speedGauge->addValues(90)->setValueRange(0,100);
-    QcLabelItem *labelType = p_speedGauge->addLabel(70);
-    labelType->setText("Mbps");
-    labelType->setColor(Qt::green);
+    p_gaugeValuesItem = p_gauge->addValues(90);
+    p_gaugeValuesItem->setValueRange(0,20);
 
-    p_upNeedle = p_speedGauge->addNeedle(70);
-    p_upNeedle->setNeedle(QcNeedleItem::NeedleType::FeatherNeedle);
-    p_upNeedle->setColor(Qt::red);
-    p_upNeedle->setValueRange(0,100);
+    p_gaugeUnitLabel = p_gauge->addLabel(70);
+    p_gaugeUnitLabel->setText("Mbps");
+    p_gaugeUnitLabel->setColor(Qt::green);
 
-    QcLabelItem *labelDown = p_speedGauge->addLabel(40);
-    labelDown->setText("0.0");
-    labelDown->setColor(Qt::green);
-    p_downNeedle = p_speedGauge->addNeedle(80);
-    p_downNeedle->setNeedle(QcNeedleItem::NeedleType::FeatherNeedle);
-    p_downNeedle->setLabel(labelDown);
-    p_downNeedle->setColor(Qt::green);
-    p_downNeedle->setValueRange(0,100);
+    p_gaugeUpNeedle = p_gauge->addNeedle(70);
+    p_gaugeUpNeedle->setNeedle(QcNeedleItem::NeedleType::FeatherNeedle);
+    p_gaugeUpNeedle->setColor(Qt::red);
+    p_gaugeUpNeedle->setValueRange(0,20);
+
+    p_gaugeSpeedLabel = p_gauge->addLabel(40);
+    p_gaugeSpeedLabel->setText("0.0");
+    p_gaugeSpeedLabel->setColor(Qt::green);
+
+    p_gaugeDownNeedle = p_gauge->addNeedle(80);
+    p_gaugeDownNeedle->setNeedle(QcNeedleItem::NeedleType::FeatherNeedle);
+    p_gaugeDownNeedle->setLabel(p_gaugeSpeedLabel);
+    p_gaugeDownNeedle->setColor(Qt::green);
+    p_gaugeDownNeedle->setValueRange(0,20);
 
 //    p_SpeedGauge->addBackground(7);
 //    p_SpeedGauge->addGlass(88);
-    ui->contentGauge->addWidget(p_speedGauge);
+    ui->contentGauge->addWidget(p_gauge);
 
 }
 
@@ -217,10 +230,22 @@ void MainWindow::readSettings()
     QSettings settings(qApp->property("iniFilePath").toString(), QSettings::IniFormat);
 
     settings.beginGroup("MainWindow");
+
     if ( settings.value("fullScreen", false).toBool() ) setWindowState(Qt::WindowFullScreen);
     resize(settings.value("size", QSize(840, 670)).toSize());
     move(settings.value("pos", QPoint(200, 200)).toPoint());
+
+    settings.beginGroup("Gauge");
+    float downValue = settings.value("maxDownloadValue", 20).toFloat();
+    float upValue = settings.value("maxUploadValue", 20).toFloat();
+    p_gaugeValuesItem->setMaxValue(downValue);
+    p_gaugeDegreesItem->setMaxValue(downValue);
+    p_gaugeDownNeedle->setMaxValue(downValue);
+    p_gaugeUpNeedle->setMaxValue(upValue);
     settings.endGroup();
+
+    settings.endGroup();
+
     qDebug() << "Settings read";
 
     bool justUpdated = settings.value("justUpdated", false).toBool();
@@ -242,6 +267,10 @@ void MainWindow::writeSettings()
     settings.setValue("fullScreen", isFullScreen());
     settings.setValue("size", size());
     settings.setValue("pos", pos());
+    settings.beginGroup("Gauge");
+    settings.setValue("maxDownloadValue", p_gaugeDownNeedle->getMaxValue());
+    settings.setValue("maxUploadValue", p_gaugeUpNeedle->getMaxValue());
+    settings.endGroup();
     settings.endGroup();
 
     settingsPage->saveSettings();
@@ -398,6 +427,29 @@ void MainWindow::listenerUpdate(QString type, bool success)
     }
 }
 
+void MainWindow::sessionStatisticUpdate(const quint64 &sent, const quint64 &received)
+{
+    QPair<float, float> pair = downloadPage->getRate();
+    float downRate = pair.first;
+    float upRate = pair.second;
+
+//    QString msg = QString("sent %0%1 / received %2%3").arg(convertSize(sent)).arg(convertSizeUnit(sent))
+//                                                        .arg(convertSize(received)).arg(convertSizeUnit(received));
+
+    QString htmlDown = "<img src='qrc:/images/arrow_down.png'></img>";
+    QString htmlUp = "<img src='qrc:/images/arrow_up.png'></img>";
+
+    QString ups = QString("%0 %1%2/s (%3%4) / %5 %6%7/s (%8%9)").arg(htmlDown).arg(convertSize(downRate)).arg(convertSizeUnit(downRate))
+                                                                              .arg(convertSize(received)).arg(convertSizeUnit(received))
+                                                                .arg(htmlUp).arg(convertSize(upRate)).arg(convertSizeUnit(upRate))
+                                                                            .arg(convertSize(sent)).arg(convertSizeUnit(sent));
+    updateStatusBarStatistics(ups);
+
+    updateGauge((downRate * 8)/1000000, (upRate * 8)/1000000);  // Ethernet 100 BASE-T -> http://www.convert-me.com/en/convert/data_transfer_rate/byte_s.html?u=byte%2Fs&v=1
+
+    statisticsPage->updateStats(pair);
+}
+
 void MainWindow::closeEvent(QCloseEvent *event)
 {
     qDebug("closing");
@@ -417,4 +469,45 @@ void MainWindow::changeEvent(QEvent *e)
     }
 
     QMainWindow::changeEvent(e);
+}
+
+QString MainWindow::convertSize(const int &size)
+{
+    if (size==0) return "0";
+    float num = size;
+    QStringList list;
+    list << "KB" << "MB" << "GB" << "TB";
+
+    QStringListIterator i(list);
+    QString unit("b");
+
+    while(num >= 1000.0 && i.hasNext())
+     {
+        unit = i.next();
+        num /= 1000.0;
+    }
+
+    int length = 1;
+    int x = num;
+    while ( x /= 10 )
+       length++;
+
+    return QString().setNum(num,'f',3-length);
+}
+
+QString MainWindow::convertSizeUnit(const int &size)
+{
+    float num = size;
+    QStringList list;
+    list << "KB" << "MB" << "GB" << "TB";
+
+    QStringListIterator i(list);
+    QString unit("b");
+
+    while(num >= 1024.0 && i.hasNext())
+     {
+        unit = i.next();
+        num /= 1024.0;
+    }
+    return unit;
 }