+++ settingswindow.h
@@ -11,6 +11,8 @@
 #include <QApplication>
 #include <QProcess>
 
+#include <changelog.h>
+
 namespace Ui {
 class settingswindow;
 }
@@ -45,6 +47,8 @@ private slots:
     void on_btnSave_released();
     void on_btnCancel_released();
     void on_btnFolder_released();
+
+    void on_pushButton_released();
 
 private:
     Ui::settingswindow *ui;